fix rsp, adjust intr entry codes

+++ kernel/include/proto.h
@@ -21,7 +21,7 @@
 	void video_init(void);
 
 	/* exception.c */
-	void exception_handler(unsigned long);
+	void exception_handler(unsigned int);
 
 	/* functions defined in architecture-dependent files. */
 	void prot_init(void);
